api-backend/backendProject/requirements.txt
Django==2.2
djangorestframework
mysqlclient
django-cors-headers
firebase-admin